requirements.txt
# TTS Agent Standalone 依赖

# Web 框架
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6

# HTTP 客户端
httpx>=0.26.0
aiohttp>=3.9.0

# 数据库
sqlalchemy>=2.0.0

# LangChain 相关
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20

# 音频处理
pydub>=0.25.1

# 数据验证
pydantic>=2.5.0

# 环境变量
python-dotenv>=1.0.0

# 工具库
uuid7>=0.1.0
